refactor: remove add videoCallData
